move buffer.hpp, chained_buffer.hpp and receive_buffer.hpp into aux_, since they are internal types
