+++ include/libtorrent/pe_crypto.hpp
@@ -41,12 +41,11 @@ POSSIBILITY OF SUCH DAMAGE.
 #include <boost/multiprecision/cpp_int.hpp>
 #include "libtorrent/aux_/disable_warnings_pop.hpp"
 
-#include "libtorrent/receive_buffer.hpp"
+#include "libtorrent/aux_/receive_buffer.hpp"
 #include "libtorrent/sha1_hash.hpp"
 #include "libtorrent/extensions.hpp"
 #include "libtorrent/assert.hpp"
 #include "libtorrent/span.hpp"
-#include "libtorrent/buffer.hpp"
 #include "libtorrent/aux_/array.hpp"
 
 #include <list>
@@ -100,14 +99,14 @@ namespace libtorrent {
 		std::tuple<int, span<span<char const>>>
 		encrypt(span<span<char>> iovec);
 
-		int decrypt(crypto_receive_buffer& recv_buffer
+		int decrypt(aux::crypto_receive_buffer& recv_buffer
 			, std::size_t& bytes_transferred);
 
 		bool switch_send_crypto(std::shared_ptr<crypto_plugin> crypto
 			, int pending_encryption);
 
 		void switch_recv_crypto(std::shared_ptr<crypto_plugin> crypto
-			, crypto_receive_buffer& recv_buffer);
+			, aux::crypto_receive_buffer& recv_buffer);
 
 		bool is_send_plaintext() const
 		{
